requirements.txt
botocore==1.29.146
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
six==1.16.0
urllib3==1.26.16
